pyproject.toml
[tool.poetry]
name = "trips-reminder-example"
version = "0.1.0"
description = "Write something about this project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
aiogram = "^3.13.1"
pydantic = "^2.9.2"
aiosqlite = "^0.20.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
celery = {extras = ["async"], version = "^5.4.0"}
flower = "^2.0.1"
redis = "^5.1.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
